backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.14.1
pydantic-settings==2.10.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
pytest==8.4.1
mongomock==4.3.0
httpx==0.28.1
